Fixed occasional error debug print macro
If debug macro was used without passing arguments to print the
format-string then one would see the following error-

error: expected expression before ')' token

This was due to the presence of the trailing comma after fmt.
Without any arguments to print, it is followed by ')'.

Modified the debug macro to support zero arguments to print.

Signed-off-by: Chinmay V S <chinmay.v.s@pathpartnertech.com>