Cleanup use of "__signed" keyword
Switch to using "signed" keyword as all modern compilers recognise it.

When __KERNEL__ is not defined then that part of the header may be
exposed to userspace[1]. Older compilers (or newer versions of gcc with
-traditional used to compile old programs) don't recognize the keyword
'signed', hence the alternate keyword is used.
Extreme backward compatibility I'd say ;) I doubt that today is
possibile to compile anything with -traditional on a distro recent
enough to use kernel 2.6.

Source: http://lkml.iu.edu/hypermail/linux/kernel/0610.3/1379.html

Signed-off-by: TheCodeArtist <cvs268@gmail.com>